Variational Bayesian inference for unsupervised clustering, mixture of categorical variables

Variational Bayesian inference for unsupervised clustering, mixture of categorical variables

vimixCat(X, K, prior, init = "kmodes", tol = 1e-04, maxiter = 2000,
  verbose = F)

Arguments

X
NxD data matrix.

K
(Maximum) number of clusters.

prior
Prior parameters (optional).

init
Initialisation method (optional). If it is a vector, it is interpreted as the vector of initial cluster allocations. If it is a string, it is interpreted as the name of the clustering algorithm used for the initialisation (only "kmeans" and "random") available at the moment).

tol
Tolerance on lower bound. Default is 10e-20.

maxiter
Maximum number of iterations of the VB algorithm. Default is 2000.

verbose
Boolean flag which, if TRUE, prints the iteration numbers. Default is FALSE.

Value

A list containing L, the lower bound at each step of the algorithm, label, a vector containing the cluster labels, model, a list containing the trained model structure.

References

Bishop, C.M., 2006. Pattern recognition and machine learning. Springer.